requirements.txt
httpx
python-jose
python-multipart
fastapi>=0.108.0
pydantic_settings
pydantic>=2.5
sqlalchemy_utils
pytz
email-validator
deepdiff
unidecode
pysql-repo>=0.6.0
loguru
opentelemetry-api
opentelemetry-exporter-otlp-proto-http
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-sqlalchemy
opentelemetry-sdk
